separated methods from fleet into vehicle, still need to add macro and fix types
