+++ src/sprint5/ITaxiCompany.java
@@ -9,8 +9,8 @@ public interface ITaxiCompany {
     public String getName();    
     public int getTotalServices();
     public boolean provideService(int user);
-    public void arrivedAtPickupLocation(IVehicle vehicle);
-    public void arrivedAtDropoffLocation(IVehicle vehicle);
+    public void arrivedAtPickupLocation(IFleet vehicle);
+    public void arrivedAtDropoffLocation(IFleet vehicle);
 
         
     }
